requirements.txt
requests>=2.28.0
squareup>=26.0.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
python-dotenv>=1.0.0
